[Core] modified structure of the code. Splitted FESTIM.py in smaller modules + fixed the tests
