+++ FESTIM/__init__.py
@@ -1,0 +1,17 @@
+from . import helpers
+
+from . import meshing
+from . import functionspaces_and_functions
+from . import initialise_solutions
+from . import formulations
+
+from . import boundary_conditions
+from . import solving
+
+from . import export
+from . import post_processing
+
+from . import generic_simulation
+
+import sympy as sp
+x, y, z, t = sp.symbols('x[0] x[1] x[2] t')
