+++ Tests/test_meshing.py
@@ -1,5 +1,5 @@
 # Unit tests meshing
-from FESTIM import FESTIM
+import FESTIM
 import fenics
 import pytest
 import sympy as sp
@@ -11,7 +11,7 @@ def test_mesh_and_refine_meets_refinement_conditions():
     refinement conditions
     '''
     def create_subdomains(x1, x2):
-        class domain(FESTIM.SubDomain):
+        class domain(fenics.SubDomain):
             def inside(self, x, on_boundary):
                 return x[0] >= x1 and x[0] <= x2
         domain = domain()
@@ -29,17 +29,17 @@ def test_mesh_and_refine_meets_refinement_conditions():
     refinements = [[[2, 3], [0.5, 0.25]], [[3, 11], [0.5, 0.25]]]
     for i in range(len(refinements)):
         param = mesh_parameters(2, 1, refinements[i][0], refinements[i][1])
-        mesh = FESTIM.mesh_and_refine(param)
+        mesh = FESTIM.meshing.mesh_and_refine(param)
 
-        mf1 = FESTIM.MeshFunction('size_t', mesh, 1)
-        mf2 = FESTIM.MeshFunction('size_t', mesh, 1)
+        mf1 = fenics.MeshFunction('size_t', mesh, 1)
+        mf2 = fenics.MeshFunction('size_t', mesh, 1)
         subdomain1 = create_subdomains(0, refinements[i][1][1])
         subdomain1.mark(mf1, 1)
         subdomain2 = create_subdomains(0, refinements[i][1][0])
         subdomain2.mark(mf2, 2)
         nb_cell_1 = 0
         nb_cell_2 = 0
-        for cell in FESTIM.cells(mesh):
+        for cell in fenics.cells(mesh):
             cell_no = cell.index()
             if mf1.array()[cell_no] == 1:
                 nb_cell_1 += 1
